+++ Frasy/src/utils/misc/deserializer.h
@@ -27,10 +27,8 @@
 #include <type_traits>
 #include <vector>
 
-namespace Frasy
-{
-namespace Internal
-{
+namespace Frasy {
+namespace Internal {
 template<typename T>
 concept Primitives = std::is_arithmetic_v<T> || std::same_as<T, std::string>;
 }
@@ -62,11 +60,10 @@ template<typename T, typename Begin, typename End>
     requires((std::is_integral_v<T> && !std::is_same_v<T, bool>))
 T Deserialize(Begin&& b, End&& e)
 {
-    if (sizeof(T) > std::distance(b, e)) { throw std::runtime_error("Not enough data!"); }
+    if (static_cast<long long int>(sizeof(T)) > std::distance(b, e)) { throw std::runtime_error("Not enough data!"); }
     constexpr size_t N = sizeof(T);
     T                t = {};
-    for (size_t i = 1; i <= sizeof(T); i++)
-    {
+    for (size_t i = 1; i <= sizeof(T); i++) {
         t |= static_cast<T>(*b) << ((N * 8) - (i * 8));
         ++b;
     }
@@ -79,8 +76,7 @@ T Deserialize(Begin&& b, End&& e)
 {
     constexpr size_t          N = sizeof(T);
     std::underlying_type_t<T> t = {};
-    for (size_t i = 1; i <= sizeof(T); i++)
-    {
+    for (size_t i = 1; i <= sizeof(T); i++) {
         t |= *b << ((N * 8) - (i * 8));
         ++b;
     }
@@ -116,24 +112,22 @@ bool Deserialize(Begin&& b, End&& e)
 template<SerializableContainer T, typename Begin, typename End>
 T Deserialize(Begin&& b, End&& e)
 {
-    if (std::distance(b, e) < sizeof(uint16_t)) { throw std::runtime_error("Not enough data"); }
+    if (std::distance(b, e) < static_cast<long long int>(sizeof(uint16_t))) {
+        throw std::runtime_error("Not enough data");
+    }
     T        t;
-    uint16_t size = Deserialize<uint16_t>(b, e);    // Get reported size from serializer
-    if constexpr (std::is_same_v<std::array<typename T::value_type, sizeof(T) / sizeof(typename T::value_type)>, T>)
-    {
+    uint16_t size = Deserialize<uint16_t>(b, e); // Get reported size from serializer
+    if constexpr (std::is_same_v<std::array<typename T::value_type, sizeof(T) / sizeof(typename T::value_type)>, T>) {
         // Array, check size is valid
-        if (size != sizeof(T) / sizeof(typename T::value_type))
-        {
+        if (size != sizeof(T) / sizeof(typename T::value_type)) {
             throw std::runtime_error("Received invalid size for array!");
         }
     }
-    else if constexpr (std::is_same_v<std::vector<typename T::value_type>, T>)
-    {
+    else if constexpr (std::is_same_v<std::vector<typename T::value_type>, T>) {
         // Vector, must resize
         t.resize(size);
     }
-    else if constexpr (std::is_same_v<std::string, T>)
-    {
+    else if constexpr (std::is_same_v<std::string, T>) {
         // String, must resize
         t.resize(size);
     }
@@ -151,7 +145,7 @@ T Deserialize(Begin&& b, End&& e)
     return t;
 }
 
-}    // namespace Frasy
+} // namespace Frasy
 
 //!@}
